+++ src/routes/groceries.js
@@ -1,7 +1,5 @@
 const {Router} = require('express');
-
 const router = Router();
-
 const groList = [
     {
         item:"milk",
@@ -20,22 +18,16 @@ const groList = [
         price:176
     },
     
-    
 ]
-
 router.get('/',(req,res)=>{
     res.send(groList)
 });
-
 router.get('/:item',(req,res)=>{
     const {item} = req.params;
-    console.log(item)
     const groItem = groList.find((g)=> g.item === item);
     res.send(groItem)
 });
-
 router.post('/',(req,res)=>{
-    console.log(req.body);
     groList.push(req.body);
     res.send(201);
 });
